Added two new fields to the OctoMap class.
The OccupiedVoxelCount field is simply a counter which stores the number of occupied voxels within the octomap. Allows for a easier way to check for the current number of occupied voxels.

The HasChanged field is a flag that is set true when a new point gets added to the OctoMap. It is then set to false when the HasChanged function gets called. Can be used to determine if the OctoMap has changed at all.